Реализация истории использования "как (пользователь, модератор, корректор) я хочу получать уведомления от топиков, теги которых совпадают с одним из тегов в моём профиле"

+++ src/main/java/ru/org/linux/user/UserTagDao.java
@@ -17,22 +17,26 @@ package ru.org.linux.user;
 
 import com.google.common.collect.ImmutableList;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowCallbackHandler;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 @Repository
 public class UserTagDao {
 
-  private JdbcTemplate jdbcTemplate;
+  private NamedParameterJdbcTemplate jdbcTemplate;
 
   @Autowired
   public void setDataSource(DataSource ds) {
-    jdbcTemplate = new JdbcTemplate(ds);
+    jdbcTemplate = new NamedParameterJdbcTemplate(ds);
   }
 
   /**
@@ -43,11 +47,13 @@ public class UserTagDao {
    * @param isFavorite выбирать фаворитные теги (true) или игнорируемые (false)
    */
   public void addTag(int userId, int tagId, boolean isFavorite) {
+    MapSqlParameterSource parameters = new MapSqlParameterSource();
+    parameters.addValue("user_id", userId);
+    parameters.addValue("tag_id", tagId);
+    parameters.addValue("is_favorite", isFavorite);
     jdbcTemplate.update(
-      "INSERT INTO user_tags (user_id, tag_id, is_favorite) VALUES(?,?,?)",
-      userId,
-      tagId,
-      isFavorite
+      "INSERT INTO user_tags (user_id, tag_id, is_favorite) VALUES(:user_id, :tag_id, :is_favorite)",
+      parameters
     );
   }
 
@@ -59,11 +65,14 @@ public class UserTagDao {
    * @param isFavorite выбирать фаворитные теги (true) или игнорируемые (false)
    */
   public void deleteTag(int userId, int tagId, boolean isFavorite) {
+    MapSqlParameterSource parameters = new MapSqlParameterSource();
+    parameters.addValue("user_id", userId);
+    parameters.addValue("tag_id", tagId);
+    parameters.addValue("is_favorite", isFavorite);
+
     jdbcTemplate.update(
-      "DELETE FROM user_tags WHERE user_id=? and tag_id=? and is_favorite=?",
-      userId,
-      tagId,
-      isFavorite
+      "DELETE FROM user_tags WHERE user_id=:user_id and tag_id=:tag_id and is_favorite=:is_favorite",
+      parameters
     );
   }
 
@@ -73,7 +82,9 @@ public class UserTagDao {
    * @param tagId идентификационный номер тега
    */
   public void deleteTags(int tagId) {
-    jdbcTemplate.update("DELETE FROM user_tags WHERE tag_id=?", tagId);
+    MapSqlParameterSource parameters = new MapSqlParameterSource();
+    parameters.addValue("tag_id", tagId);
+    jdbcTemplate.update("DELETE FROM user_tags WHERE tag_id=:tag_id", parameters);
   }
 
   /**
@@ -84,23 +95,40 @@ public class UserTagDao {
    * @return список тегов пользователя
    */
   public ImmutableList<String> getTags(int userId, boolean isFavorite) {
+    MapSqlParameterSource parameters = new MapSqlParameterSource();
+    parameters.addValue("user_id", userId);
+    parameters.addValue("is_favorite", isFavorite);
+
     final ImmutableList.Builder<String> tags = ImmutableList.builder();
 
     jdbcTemplate.query(
       "SELECT tags_values.value FROM user_tags, tags_values WHERE " +
-        "user_tags.user_id=? AND tags_values.id=user_tags.tag_id AND user_tags.is_favorite=?" +
+        "user_tags.user_id=:user_id AND tags_values.id=user_tags.tag_id AND user_tags.is_favorite=:is_favorite " +
         "ORDER BY value",
+      parameters,
       new RowCallbackHandler() {
         @Override
         public void processRow(ResultSet rs) throws SQLException {
           tags.add(rs.getString("value"));
         }
-      },
-      userId,
-      isFavorite
+      }
     );
 
     return tags.build();
   }
 
+  /**
+   * Получить список ID пользователей, у которых в профиле есть перечисленные фаворитные теги.
+   *
+   * @param tags  список фаворитных тегов
+   * @return список ID пользователей
+   */
+  public List<Integer> getUserIdListByTags(List<String> tags) {
+    return jdbcTemplate.queryForList(
+      "select distinct user_id from user_tags where tag_id in (select id from tags_values where value in ( :values )) "
+        + "AND is_favorite = true",
+      Collections.singletonMap("values", tags),
+      Integer.class
+    );
+  }
 }
